LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/sameDocument-after-navigate-restore-expected.txt



PASS entry.sameDocument is properly restored after cross-document back
